controllers/sauce.js: modif ctrler del + mid/auth

+++ backend/controllers/sauce.js
@@ -34,10 +34,27 @@ exports.modifySauce = (req, res, next) => {
 
 //Controller DELETE
 exports.deleteSauce = (req, res, next) => {
-    //Supprimer une sauce de la base de données, en fonction de l'id
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-        .catch(error => res.status(400).json({ error }))
+    // S'assurer que l'utilisateur qui fait la requête est bien le propriétaire de l'objet qu'il souhaite supprimer
+    Thing.findOne({ _id: req.params.id }).then(
+        (sauce) => {
+            // Cas d'erreur: si sauce n'existe pas
+            if (!sauce) {
+                return res.status(404).json({
+                    error: new Error('Objet non trouvé !')
+                })
+            }
+            // Vérification du userId 
+            if (sauce.userId !== req.auth.userId) {
+                return res.status(401).json({
+                    error: new Error('Requête non autorisée !')
+                })
+            }
+            //Si bon propriétaire: supprimer une sauce de la base de données, en fonction de l'id
+            Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
+            .catch(error => res.status(400).json({ error }))
+        }
+    )
 }
